seperate laws app tests to different files and surprise! win another 70 tests that are not run usually
